feat: make-zksync-airdrop-modal-globally
