+++ apps/web/src/views/Home/components/Banners/ZksyncAirdropBanner.tsx
@@ -15,39 +15,10 @@ import {
   VerticalDivider,
 } from '@pancakeswap/widgets-internal'
 import { ClaimZksyncAirdropModal } from 'components/ClaimZksyncAirdropModal'
-import { useUserWhiteListData, useZksyncAirDropData } from 'components/ClaimZksyncAirdropModal/hooks'
 import { ASSET_CDN } from 'config/constants/endpoints'
-import { useAtom } from 'jotai'
-import { atomWithStorage } from 'jotai/utils'
 import Image from 'next/legacy/image'
-import { useCallback, useEffect, useRef, useState } from 'react'
+import { useCallback, useState } from 'react'
 import styled from 'styled-components'
-import { useAccount } from 'wagmi'
-
-const zksyncAutoPopup = atomWithStorage<{ [account: string]: boolean }>(
-  'pcs:zksync-airdrop-auto-popup-v2',
-  {},
-  undefined,
-  { unstable_getOnInit: true },
-)
-
-const useAutoPopup = (): [boolean, (value: boolean) => void] => {
-  const { address } = useAccount()
-  const [autoPopup, setAutoPopup] = useAtom(zksyncAutoPopup)
-
-  const setPopupState = useCallback(
-    (value: boolean) => {
-      if (!address) return
-      setAutoPopup((prev) => ({
-        ...prev,
-        [address]: value,
-      }))
-    },
-    [address, setAutoPopup],
-  )
-
-  return [address ? autoPopup[address] ?? true : false, setPopupState]
-}
 
 const bgMobile = `${ASSET_CDN}/web/banners/zksync-airdrop-banner/bunny-bg-mobile.png`
 const bgDesktop = `${ASSET_CDN}/web/banners/zksync-airdrop-banner/bunny-bg.png`
@@ -84,30 +55,11 @@ export const ZksyncAirDropBanner = () => {
   const { t } = useTranslation()
   const { isMobile, isTablet, isXs } = useMatchBreakpoints()
   const [isOpen, setIsOpen] = useState(false)
-  const [autoPopup, setAutoPopup] = useAutoPopup()
-  const whitelistData = useUserWhiteListData(autoPopup || isOpen)
-  const { zksyncAirdropData } = useZksyncAirDropData(whitelistData?.proof)
-  const isModalOpened = useRef(false)
 
-  useEffect(() => {
-    if (whitelistData?.account) {
-      if (!whitelistData?.amount && !whitelistData?.proof) {
-        setAutoPopup(false)
-      } else if (
-        whitelistData?.amount &&
-        whitelistData?.proof &&
-        isModalOpened.current === false &&
-        autoPopup &&
-        zksyncAirdropData?.claimedAmount === 0n
-      ) {
-        setIsOpen(true)
-        isModalOpened.current = true
-        setAutoPopup(false)
-      }
-    }
-  }, [whitelistData, zksyncAirdropData, autoPopup, setAutoPopup])
-
-  const handleDismiss = useCallback(() => {
+  const onOpen = useCallback(() => {
+    setIsOpen(true)
+  }, [])
+  const onDismiss = useCallback(() => {
     setIsOpen(false)
   }, [])
 
@@ -138,13 +90,7 @@ export const ZksyncAirDropBanner = () => {
           desc={<BannerDesc color="white">{!isMobile ? t('2.4 Million $ZK available to claim') : null}</BannerDesc>}
           actions={
             <BannerActionContainer>
-              <LinkExternalAction
-                onClick={() => {
-                  setIsOpen(true)
-                }}
-                color="#FFE238"
-                externalIcon="arrowForward"
-              >
+              <LinkExternalAction onClick={onOpen} color="#FFE238" externalIcon="arrowForward">
                 {t('Check Eligibility')}
               </LinkExternalAction>
               {!isMobile && (
@@ -167,7 +113,7 @@ export const ZksyncAirDropBanner = () => {
           <FloatingGraphic src={floatingAsset} width={isMobile ? 100 : 80} height={isMobile ? 100 : 80} />
         </BannerGraphics>
       </BannerContainer>
-      <ClaimZksyncAirdropModal isOpen={isOpen} onDismiss={handleDismiss} />
+      <ClaimZksyncAirdropModal isOpen={isOpen} onDismiss={onDismiss} />
     </>
   )
 }
